// src/app/core/models/enums/role.enum.ts
export enum EmployeeRole {
  EMPLEADO_BASICO = 'empleado_basico',
  EMPLEADO = 'empleado',
  RECEPCIONISTA = 'recepcionista',
  GERENTE = 'gerente',
  PERSONALIZADO = 'personalizado'
}

export enum RoleValues {
  EMPLEADO_BASICO = '{"name":"Empleado básico","value":"empleado_basico"}',
  EMPLEADO = '{"name":"Empleado","value":"empleado"}',
  RECEPCIONISTA = '{"name":"Recepcionista","value":"recepcionista"}',
  GERENTE = '{"name":"Gerente","value":"gerente"}',
  PERSONALIZADO = '{"name":"Personalizado","value":"personalizado"}'
}